Break out filterable toolbar into it's own component.
