+++ src/Containers/JobExplorer/JobExplorer.js
@@ -24,48 +24,17 @@ import {
 } from '@redhat-cloud-services/frontend-components';
 
 import {
-    FilterIcon,
-    CalendarAltIcon,
-    QuestionCircleIcon
-} from '@patternfly/react-icons';
-
-import {
-    Toolbar,
-    ToolbarContent as PFToolbarContent,
-    ToolbarFilter,
-    ToolbarToggleGroup,
-    ToolbarItem,
-    ToolbarGroup as PFToolbarGroup,
     Badge,
     Card,
     CardBody,
     CardHeader as PFCardHeader,
     Pagination,
-    PaginationVariant,
-    Select,
-    SelectOption,
-    // SelectVariant,
-    // DropdownPosition,
-    Dropdown,
-    DropdownToggle,
-    DropdownItem,
-    InputGroup,
-    InputGroupText,
-    TextInput,
-    Switch as PFSwitch,
-    Tooltip,
-    TooltipPosition
+    PaginationVariant
 } from '@patternfly/react-core';
 
 import JobExplorerList from '../../Components/JobExplorerList';
+import FilterableToolbar from '../../Components/Toolbar';
 
-const ToolbarGroup = styled(PFToolbarGroup)`
-  button {
-    .pf-c-select__toggle-wrapper {
-      flex-wrap: nowrap;
-    }
-  }
-`;
 const CardHeader = styled(PFCardHeader)`
   display: flex;
   justify-content: space-between;
@@ -82,18 +51,6 @@ const TitleWithBadge = styled.div`
   h2 {
     margin-right: 10px;
   }
-`;
-
-const Switch = styled(PFSwitch)`
-    &&& {
-        margin: 0 15px;
-    }
-`;
-
-const ToolbarContent = styled(PFToolbarContent)`
-    .pf-c-toolbar__content-section {
-        justify-content: space-between;
-    }
 `;
 
 const perPageOptions = [
@@ -120,16 +77,6 @@ const JobExplorer = props => {
     const [ isLoading, setIsLoading ] = useState(true);
     const [ meta, setMeta ] = useState({});
     const [ currPage, setCurrPage ] = useState(1);
-    const [ statusIsExpanded, setStatusIsExpanded ] = useState(false);
-    const [ dateRangeIsExpanded, setDateRangeIsExpanded ] = useState(false);
-    const [ jobTypeIsExpanded, setJobTypeIsExpanded ] = useState(false);
-    const [ orgIsExpanded, setOrgIsExpanded ] = useState(false);
-    const [ isCategoryExpanded, setIsCategoryExpanded ] = useState(false);
-    const [ clusterIsExpanded, setClusterIsExpanded ] = useState(false);
-    const [ templateIsExpanded, setTemplateIsExpanded ] = useState(false);
-    const [ sortByIsExpanded, setSortByIsExpanded ] = useState(false);
-    const [ currentCategory, setCurrentCategory ] = useState('Status');
-
     const [ orgIds, setOrgIds ] = useState([]);
     const [ clusterIds, setClusterIds ] = useState([]);
     const [ templateIds, setTemplateIds ] = useState([]);
@@ -317,32 +264,6 @@ const JobExplorer = props => {
         return () => (ignore = true);
     }, []);
 
-    const handleChips = (item, comparator) => {
-        return item.reduce((acc, i) => {
-            Number.isInteger(parseInt(i)) ? (i = parseInt(i)) : i;
-            comparator.forEach(item => {
-                if (item.key === i) {
-                    acc.push(item.value);
-                }
-            });
-            return acc;
-        }, []);
-    };
-
-    const handleDateChips = (date, comparator) => {
-        if (date && typeof date === 'string') {
-            let val;
-            comparator.forEach(i => {
-                if (i.key === date) {
-                    val = i.value;
-                }
-            });
-            return new Array(val);
-        }
-
-        return new Array();
-    };
-
     const onDelete = (type, val) => {
         let filtered;
         Number.isInteger(val) ? (val = parseInt(val)) : val;
@@ -403,269 +324,6 @@ const JobExplorer = props => {
         }
     };
 
-    const buildCategoryDropdown = () => {
-        return (
-            <ToolbarItem>
-                <Dropdown
-                    onSelect={ e => {
-                        setCurrentCategory(e.target.innerText);
-                        setIsCategoryExpanded(!isCategoryExpanded);
-                    } }
-                    position={ 'left' }
-                    toggle={
-                        <DropdownToggle
-                            onToggle={ () => {
-                                setIsCategoryExpanded(!isCategoryExpanded);
-                            } }
-                            style={ { width: '100%' } }
-                        >
-                            <FilterIcon />
-              &nbsp;{ currentCategory }
-                        </DropdownToggle>
-                    }
-                    isOpen={ isCategoryExpanded }
-                    dropdownItems={ [
-                        <DropdownItem key="cat0">Status</DropdownItem>,
-                        <DropdownItem key="cat6">Date</DropdownItem>,
-                        <DropdownItem key="cat1">Job type</DropdownItem>,
-                        <DropdownItem key="cat2">Organization</DropdownItem>,
-                        <DropdownItem key="cat3">Cluster</DropdownItem>,
-                        <DropdownItem key="cat4">Template</DropdownItem>,
-                        <DropdownItem key="cat5">Sort by</DropdownItem>
-                    ] }
-                    style={ { width: '100%' } }
-                />
-            </ToolbarItem>
-        );
-    };
-
-    const buildFilterDropdown = () => {
-        const organizationIdMenuItems = orgIds.map(({ key, value }) => (
-            <SelectOption key={ key } value={ `${key}` }>
-                { value }
-            </SelectOption>
-        ));
-
-        const statusMenuItems = statuses.map(({ key, value }) => (
-            <SelectOption key={ key } value={ key }>
-                { value }
-            </SelectOption>
-        ));
-
-        const jobTypeMenuItems = jobTypes.map(({ key, value }) => (
-            <SelectOption key={ key } value={ key }>
-                { value }
-            </SelectOption>
-        ));
-
-        const clusterIdMenuItems = clusterIds.map(({ key, value }) => (
-            <SelectOption key={ key } value={ `${key}` }>
-                { value }
-            </SelectOption>
-        ));
-
-        const templateIdMenuItems = templateIds.map(({ key, value }) => (
-            <SelectOption key={ key } value={ `${key}` }>
-                { value }
-            </SelectOption>
-        ));
-
-        const sortByMenuItems = sortBy.map(({ key, value }) => (
-            <SelectOption key={ key } value={ key }>
-                { value }
-            </SelectOption>
-        ));
-
-        const dateRangeMenuItems = quickDateRanges.map(({ key, value }) => (
-            <SelectOption key={ key } value={ key }>
-                { value }
-            </SelectOption>
-        ));
-
-        const onSelect = (type, event, selection) => {
-            const checked = event.target.checked;
-
-            setFilters({
-                ...filters,
-                [type]: checked
-                    ? [ ...filters[type], selection ]
-                    : filters[type].filter(value => value !== selection)
-            });
-        };
-
-        const onDateSelect = (_event, selection) => {
-            setFilters({
-                ...filters,
-                date: selection
-            });
-            setDateRangeIsExpanded(!dateRangeIsExpanded);
-        };
-
-        return (
-            <React.Fragment>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Status' }
-                    chips={ handleChips(filters.status, statuses) }
-                    categoryName="Status"
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ statusIsExpanded }
-                        variant={ 'checkbox' }
-                        aria-label="Status"
-                        onToggle={ () => {
-                            setStatusIsExpanded(!statusIsExpanded);
-                        } }
-                        onSelect={ (event, selection) => {
-                            onSelect('status', event, selection);
-                        } }
-                        selections={ filters.status }
-                        isExpanded={ statusIsExpanded }
-                        placeholderText="Filter by job status"
-                    >
-                        { statusMenuItems }
-                    </Select>
-                </ToolbarFilter>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Date' }
-                    categoryName="Date"
-                    chips={ handleDateChips(filters.date, quickDateRanges) }
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ dateRangeIsExpanded }
-                        variant={ 'single' }
-                        aria-label="Date"
-                        onToggle={ () => {
-                            setDateRangeIsExpanded(!dateRangeIsExpanded);
-                        } }
-                        onSelect={ (event, selection) => {
-                            onDateSelect(event, selection);
-                        } }
-                        selections={ filters.date }
-                        isExpanded={ dateRangeIsExpanded }
-                        placeholderText="Filter by date range"
-                    >
-                        { dateRangeMenuItems }
-                    </Select>
-                </ToolbarFilter>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Job type' }
-                    categoryName="Type"
-                    chips={ handleChips(filters.type, jobTypes) }
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ jobTypeIsExpanded }
-                        aria-label="Type"
-                        variant={ 'checkbox' }
-                        onToggle={ () => {
-                            setJobTypeIsExpanded(!jobTypeIsExpanded);
-                        } }
-                        onSelect={ (event, selection) => {
-                            onSelect('type', event, selection);
-                        } }
-                        selections={ filters.type }
-                        isExpanded={ jobTypeIsExpanded }
-                        placeholderText="Filter by job type"
-                    >
-                        { jobTypeMenuItems }
-                    </Select>
-                </ToolbarFilter>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Organization' }
-                    categoryName="Org"
-                    chips={ handleChips(filters.org, orgIds) }
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ orgIsExpanded }
-                        aria-label="Filter by Org"
-                        variant={ 'checkbox' }
-                        onToggle={ () => {
-                            setOrgIsExpanded(!orgIsExpanded);
-                        } }
-                        onSelect={ (event, selection) => {
-                            onSelect('org', event, selection);
-                        } }
-                        selections={ filters.org }
-                        isExpanded={ orgIsExpanded }
-                        placeholderText="Filter by organization"
-                    >
-                        { organizationIdMenuItems }
-                    </Select>
-                </ToolbarFilter>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Cluster' }
-                    categoryName="Cluster"
-                    chips={ handleChips(filters.cluster, clusterIds) }
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ clusterIsExpanded }
-                        aria-label="Filter by Cluster"
-                        variant={ 'checkbox' }
-                        onToggle={ () => {
-                            setClusterIsExpanded(!clusterIsExpanded);
-                        } }
-                        onSelect={ (event, selection) => {
-                            onSelect('cluster', event, selection);
-                        } }
-                        selections={ filters.cluster }
-                        isExpanded={ clusterIsExpanded }
-                        placeholderText="Filter by cluster"
-                    >
-                        { clusterIdMenuItems }
-                    </Select>
-                </ToolbarFilter>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Template' }
-                    categoryName="Template"
-                    chips={ handleChips(filters.template, templateIds) }
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ templateIsExpanded }
-                        aria-label="Filter by template"
-                        variant={ 'checkbox' }
-                        onToggle={ () => setTemplateIsExpanded(!templateIsExpanded) }
-                        onSelect={ (event, selection) => {
-                            onSelect('template', event, selection);
-                        } }
-                        selections={ filters.template }
-                        isExpanded={ templateIsExpanded }
-                        placeholderText="Filter by template"
-                    >
-                        { templateIdMenuItems }
-                    </Select>
-                </ToolbarFilter>
-                <ToolbarFilter
-                    showToolbarItem={ currentCategory === 'Sort by' }
-                    categoryName="SortBy"
-                    chips={ handleChips(filters.sortby, sortBy) }
-                    deleteChip={ onDelete }
-                >
-                    <Select
-                        isOpen={ sortByIsExpanded }
-                        aria-label="Sort by"
-                        variant={ 'checkbox' }
-                        onToggle={ () => {
-                            setSortByIsExpanded(!sortByIsExpanded);
-                        } }
-                        onSelect={ (event, selection) => {
-                            onSelect('sortby', event, selection);
-                        } }
-                        selections={ filters.sortby }
-                        isExpanded={ sortByIsExpanded }
-                        placeholderText="Sort by attribute"
-                    >
-                        { sortByMenuItems }
-                    </Select>
-                </ToolbarFilter>
-            </React.Fragment>
-        );
-    };
-
     const returnOffsetVal = page => {
         let offsetVal = (page - 1) * queryParams.limit;
         return offsetVal;
@@ -694,20 +352,6 @@ const JobExplorer = props => {
         setCurrPage(page);
     };
 
-    const handleStartDate = e => {
-        setFilters({
-            ...filters,
-            startDate: e
-        });
-    };
-
-    const handleEndDate = e => {
-        setFilters({
-            ...filters,
-            endDate: e
-        });
-    };
-
     return (
         <React.Fragment>
             <PageHeader>
@@ -733,88 +377,18 @@ const JobExplorer = props => {
                       </TitleWithBadge>
                   </CardHeader>
                   <CardBody>
-                      <Toolbar
-                          id="data-toolbar-with-chip-groups"
-                          clearAllFilters={ onDelete }
-                          collapseListedFiltersBreakpoint="xl"
-                      >
-                          <ToolbarContent>
-                              <ToolbarToggleGroup
-                                  toggleIcon={ <FilterIcon /> }
-                                  breakpoint="xl"
-                              >
-                                  { quickDateRanges.length > 0 && (
-                                      <ToolbarGroup variant="filter-group">
-                                          { buildCategoryDropdown() }
-                                          { buildFilterDropdown() }
-                                          { filters.date === 'custom' && (
-                            <>
-                              <InputGroup>
-                                  <InputGroupText
-                                      component="label"
-                                      htmlFor="startDate"
-                                  >
-                                      <CalendarAltIcon />
-                                  </InputGroupText>
-                                  <TextInput
-                                      name="startDate"
-                                      id="startDate"
-                                      type="date"
-                                      aria-label="Start Date"
-                                      value={ filters.startDate }
-                                      onChange={ e => handleStartDate(e) }
-                                  />
-                              </InputGroup>
-                              <InputGroup>
-                                  <InputGroupText
-                                      component="label"
-                                      htmlFor="endDate"
-                                  >
-                                      <CalendarAltIcon />
-                                  </InputGroupText>
-                                  <TextInput
-                                      name="endDate"
-                                      id="endDate"
-                                      type="date"
-                                      aria-label="End Date"
-                                      value={ filters.endDate }
-                                      onChange={ e => handleEndDate(e) }
-                                  />
-                              </InputGroup>
-                            </>
-                                          ) }
-                                      </ToolbarGroup>
-                                  ) }
-                              </ToolbarToggleGroup>
-                              <div>
-                                  <Switch
-                                      id="showRootWorkflowJobs"
-                                      label="Ignore nested workflows and jobs"
-                                      labelOff="Ignore nested workflows and jobs"
-                                      isChecked={ filters.showRootWorkflows }
-                                      onChange={ val => {
-                                          setFilters({
-                                              ...filters,
-                                              showRootWorkflows: val
-                                          });
-                                      } }
-                                  />
-                                  <Tooltip
-                                      position={ TooltipPosition.top }
-                                      content={
-                                          <div>
-                                              { ' ' }
-                                    If enabled, nested workflows and jobs
-                                    will not be included in the overall totals.
-                                    Enable this option to filter out duplicate entries.
-                                          </div>
-                                      }
-                                  >
-                                      <QuestionCircleIcon />
-                                  </Tooltip>
-                              </div>
-                          </ToolbarContent>
-                      </Toolbar>
+                      <FilterableToolbar
+                          orgs={ orgIds }
+                          statuses={ statuses }
+                          clusters={ clusterIds }
+                          templates={ templateIds }
+                          types={ jobTypes }
+                          sortables={ sortBy }
+                          dateRanges={ quickDateRanges }
+                          onDelete={ onDelete }
+                          passedFilters={ filters }
+                          handleFilters={ setFilters }
+                      />
                       { isLoading && <LoadingState /> }
                       { !isLoading && jobExplorerData.length <= 0 && <NoResults /> }
                       { !isLoading && jobExplorerData.length > 0 && (
