+++ src/Utilities/constants.js
@@ -12,3 +12,13 @@ export const jobExplorer = {
         'most_failed_tasks'
     ]
 };
+
+export const toolbarCategories = [
+    { name: 'Status', id: 1 },
+    { name: 'Date', id: 2 },
+    { name: 'Job', id: 3 },
+    { name: 'Organization', id: 4 },
+    { name: 'Cluster', id: 5 },
+    { name: 'Template', id: 6 },
+    { name: 'Sort by', id: 7 }
+];
